+++ src/Thoth.Json/Decode.fs
@@ -433,14 +433,12 @@ module Decode =
             if Helpers.isNull value then
                 Ok None
             else
-                // TODO: Review, is this OK?
                 match d1 path value with
                 | Ok v -> Ok (Some v)
                 | Error (_, BadField _ )
                 | Error (_, BadType (_, null))
                 | Error (_, BadPrimitive (_, null)) -> Ok None
                 | Error error -> Error error
-
 
     let oneOf (decoders : Decoder<'value> list) : Decoder<'value> =
         fun path value ->
@@ -625,10 +623,12 @@ module Decode =
     type IRequiredGetter =
         abstract Field : string -> Decoder<'a> -> 'a
         abstract At : List<string> -> Decoder<'a> -> 'a
+        abstract Raw : Decoder<'a> -> 'a
 
     type IOptionalGetter =
         abstract Field : string -> Decoder<'a> -> 'a option
         abstract At : List<string> -> Decoder<'a> -> 'a option
+        abstract Raw : Decoder<'a> -> 'a option
 
     type IGetters =
         abstract Required: IRequiredGetter
@@ -646,6 +646,11 @@ module Decode =
                                 raise (DecoderException error)
                         member __.At (fieldNames : string list) (decoder : Decoder<_>) =
                             match decodeValueError path (at fieldNames decoder) v with
+                            | Ok v -> v
+                            | Error error ->
+                                raise (DecoderException error)
+                        member __.Raw (decoder: Decoder<_>) =
+                            match decodeValueError path decoder v with
                             | Ok v -> v
                             | Error error ->
                                 raise (DecoderException error) }
@@ -670,7 +675,14 @@ module Decode =
                                 | Error error ->
                                     raise (DecoderException error)
                             else
-                                raise (DecoderException (path, BadType ("an object", v))) }
+                                raise (DecoderException (path, BadType ("an object", v)))
+                        member __.Raw (decoder: Decoder<_>) =
+                            match decodeValueError path decoder v with
+                            | Ok v -> Some v
+                            | Error (_, BadField _ )
+                            | Error (_, BadPrimitive (_, null)) -> None
+                            | Error error ->
+                                raise (DecoderException error) }
             } |> Ok
 
     ///////////////////////
@@ -834,7 +846,18 @@ module Decode =
     // Reflection ///
     ////////////////
 
-    open Microsoft.FSharp.Reflection
+    open System.Reflection
+    open FSharp.Reflection
+
+    type FieldType =
+        | Optional
+        | Required
+
+        member this.ToBool
+            with get () =
+                match this with
+                | Optional -> true
+                | Required -> false
 
     // TODO: Same API as for Thot.Json.Net.Decoder.BoxedDecoder
     type private BoxedDecoder = Decoder<obj>
@@ -849,17 +872,21 @@ module Decode =
     // This is used to force Fable use a generic comparer for map keys
     let private toMap<'key, 'value when 'key: comparison> (xs: ('key*'value) seq) = Map.ofSeq xs
 
-    let private autoObject (decoders: (string * BoxedDecoder)[]) (path : string) (value: obj) =
+    let private autoObject (decoderInfos: (FieldType * string * BoxedDecoder)[]) (path : string) (value: obj) =
         if not (Helpers.isObject value) then
             (path, BadPrimitive ("an object", value)) |> Error
         else
-            (decoders, Ok []) ||> Array.foldBack (fun (name, decoder) acc ->
+            (decoderInfos, Ok []) ||> Array.foldBack (fun (fieldType, name, decoder) acc ->
                 match acc with
                 | Error _ -> acc
                 | Ok result ->
-                    // TODO!!! Optional types shouldn't be required
-                    field name decoder path value
-                    |> Result.map (fun v -> v::result))
+                    match fieldType with
+                    | FieldType.Optional ->
+                        optional name decoder path value
+                        |> Result.map (fun v -> box v::result)
+                    | FieldType.Required ->
+                        field name decoder path value
+                        |> Result.map (fun v -> v::result))
 
     let private mixedArray msg (decoders: BoxedDecoder[]) (path: string) (values: obj[]): Result<obj list, DecoderError> =
         if decoders.Length <> values.Length then
@@ -873,31 +900,44 @@ module Decode =
                 | Ok result -> decoder path value |> Result.map (fun v -> v::result))
 
     let rec private makeUnion t isCamelCase name (path : string) (values: obj[]) =
-        match FSharpType.GetUnionCases(t) |> Array.tryFind (fun x -> x.Name = name) with
+        // The flags don't do anything in Fable, but we added for symmetry with .NET decoders
+        match FSharpType.GetUnionCases(t, BindingFlags.Public ||| BindingFlags.NonPublic) |> Array.tryFind (fun x -> x.Name = name) with
         | None -> (path, FailMessage("Cannot find case " + name + " in " + t.FullName)) |> Error
         | Some uci ->
             if values.Length = 0 then
-                FSharpValue.MakeUnion(uci, [||]) |> Ok
+                FSharpValue.MakeUnion(uci, [||], BindingFlags.Public ||| BindingFlags.NonPublic) |> Ok
             else
-                let decoders = uci.GetFields() |> Array.map (fun fi -> autoDecoder isCamelCase fi.PropertyType)
+                let decoders = uci.GetFields() |> Array.map (fun fi -> autoDecoder isCamelCase false fi.PropertyType)
                 mixedArray "union fields" decoders path values
-                |> Result.map (fun values -> FSharpValue.MakeUnion(uci, List.toArray values))
+                |> Result.map (fun values -> FSharpValue.MakeUnion(uci, List.toArray values, BindingFlags.Public ||| BindingFlags.NonPublic))
 
-    and private autoDecodeRecordsAndUnions (t: System.Type) (isCamelCase : bool) : BoxedDecoder =
-        if FSharpType.IsRecord(t) then
+    and private autoDecodeRecordsAndUnions (t: System.Type) (isCamelCase : bool) (isOptional : bool) : BoxedDecoder =
+        if FSharpType.IsRecord(t, BindingFlags.Public ||| BindingFlags.NonPublic) then
             fun path value ->
                 let decoders =
-                    FSharpType.GetRecordFields(t)
+                    FSharpType.GetRecordFields(t, BindingFlags.Public ||| BindingFlags.NonPublic)
                     |> Array.map (fun fi ->
                         let name =
                             if isCamelCase then
                                 fi.Name.[..0].ToLowerInvariant() + fi.Name.[1..]
                             else
                                 fi.Name
-                        name, autoDecoder isCamelCase fi.PropertyType)
+
+                        let fieldType, propertyType =
+                            if fi.PropertyType.IsGenericType then
+                                let fullname = fi.PropertyType.GetGenericTypeDefinition().FullName
+                                if fullname = typedefof<obj option>.FullName then
+                                    FieldType.Optional, fi.PropertyType.GenericTypeArguments.[0]
+                                else
+                                    FieldType.Required, fi.PropertyType
+                            else
+                                FieldType.Required, fi.PropertyType
+
+                        fieldType, name, autoDecoder isCamelCase fieldType.ToBool propertyType)
                 autoObject decoders path value
-                |> Result.map (fun xs -> FSharpValue.MakeRecord(t, List.toArray xs))
-        elif FSharpType.IsUnion(t) then
+                |> Result.map (fun xs -> FSharpValue.MakeRecord(t, List.toArray xs, BindingFlags.Public ||| BindingFlags.NonPublic))
+
+        elif FSharpType.IsUnion(t, BindingFlags.NonPublic) then
             fun path (value: obj) ->
                 if Helpers.isString(value) then
                     let name = Helpers.asString value
@@ -907,16 +947,23 @@ module Decode =
                     let name = Helpers.asString values.[0]
                     makeUnion t isCamelCase name path values.[1..]
                 else (path, BadPrimitive("a string or array", value)) |> Error
-        else
-            failwithf "Class types cannot be automatically deserialized: %s" t.FullName
 
-    and private autoDecoder isCamelCase (t: System.Type) : BoxedDecoder =
+        else
+            fun path _ ->
+                if isOptional then
+                    (path, BadPrimitive ("Generating an error message as the field is optional so the `option` decoders will return `None` instead of failing", null))
+                    |> Error
+                else
+                    (path, sprintf "Class types cannot be automatically deserialized: %s" t.FullName
+                    |> FailMessage) |> Error
+
+    and private autoDecoder isCamelCase (isOptional : bool) (t: System.Type)  : BoxedDecoder =
         if t.IsArray then
-            let decoder = t.GetElementType() |> autoDecoder isCamelCase
+            let decoder = t.GetElementType() |> autoDecoder isCamelCase false
             array decoder |> boxDecoder
         elif t.IsGenericType then
             if FSharpType.IsTuple(t) then
-                let decoders = FSharpType.GetTupleElements(t) |> Array.map (autoDecoder isCamelCase)
+                let decoders = FSharpType.GetTupleElements(t) |> Array.map (autoDecoder isCamelCase false)
                 fun path value ->
                     if Helpers.isArray value then
                         mixedArray "tuple elements" decoders path (Helpers.asArray value)
@@ -924,53 +971,53 @@ module Decode =
                     else (path, BadPrimitive ("an array", value)) |> Error
             else
                 let fullname = t.GetGenericTypeDefinition().FullName
-                if fullname = typedefof<obj option>.FullName
-                then t.GenericTypeArguments.[0] |> (autoDecoder isCamelCase) |> option |> boxDecoder
-                elif fullname = typedefof<obj list>.FullName
-                then t.GenericTypeArguments.[0] |> (autoDecoder isCamelCase) |> list |> boxDecoder
-                elif fullname = typedefof< Map<string, obj> >.FullName
-                then
-                    let decoder1 = t.GenericTypeArguments.[0] |> autoDecoder isCamelCase
-                    let decoder2 = t.GenericTypeArguments.[1] |> autoDecoder isCamelCase
+                if fullname = typedefof<obj option>.FullName then
+                    t.GenericTypeArguments.[0] |> (autoDecoder isCamelCase true) |> option |> boxDecoder
+                elif fullname = typedefof<obj list>.FullName then
+                    t.GenericTypeArguments.[0] |> (autoDecoder isCamelCase false) |> list |> boxDecoder
+                elif fullname = typedefof< Map<string, obj> >.FullName then
+                    let decoder1 = t.GenericTypeArguments.[0] |> autoDecoder isCamelCase false
+                    let decoder2 = t.GenericTypeArguments.[1] |> autoDecoder isCamelCase false
                     fun path value ->
                         match array (tuple2 decoder1 decoder2) path value with
                         | Error er -> Error er
                         | Ok ar -> toMap (unbox ar) |> box |> Ok
-                else autoDecodeRecordsAndUnions t isCamelCase
+                else
+                    autoDecodeRecordsAndUnions t isCamelCase isOptional
         else
             let fullname = t.FullName
-            if fullname = typeof<bool>.FullName
-            then boxDecoder bool
-            elif fullname = typeof<string>.FullName
-            then boxDecoder string
-            elif fullname = typeof<int>.FullName
-            then boxDecoder int
-            elif fullname = typeof<float>.FullName
-            then boxDecoder float
-            elif fullname = typeof<decimal>.FullName
-            then boxDecoder decimal
-            elif fullname = typeof<int64>.FullName
-            then boxDecoder int64
-            elif fullname = typeof<uint32>.FullName
-            then boxDecoder uint32
-            elif fullname = typeof<uint64>.FullName
-            then boxDecoder uint64
-            elif fullname = typeof<bigint>.FullName
-            then boxDecoder bigint
-            elif fullname = typeof<System.DateTime>.FullName
-            then boxDecoder datetime
-            elif fullname = typeof<System.DateTimeOffset>.FullName
-            then boxDecoder datetimeOffset
-            elif fullname = typeof<System.Guid>.FullName
-            then boxDecoder guid
-            elif fullname = typeof<obj>.FullName
-            then value
-            else autoDecodeRecordsAndUnions t isCamelCase
+            if fullname = typeof<bool>.FullName then
+                boxDecoder bool
+            elif fullname = typeof<string>.FullName then
+                boxDecoder string
+            elif fullname = typeof<int>.FullName then
+                boxDecoder int
+            elif fullname = typeof<float>.FullName then
+                boxDecoder float
+            elif fullname = typeof<decimal>.FullName then
+                boxDecoder decimal
+            elif fullname = typeof<int64>.FullName then
+                boxDecoder int64
+            elif fullname = typeof<uint32>.FullName then
+                boxDecoder uint32
+            elif fullname = typeof<uint64>.FullName then
+                boxDecoder uint64
+            elif fullname = typeof<bigint>.FullName then
+                boxDecoder bigint
+            elif fullname = typeof<System.DateTime>.FullName then
+                boxDecoder datetime
+            elif fullname = typeof<System.DateTimeOffset>.FullName then
+                boxDecoder datetimeOffset
+            elif fullname = typeof<System.Guid>.FullName then
+                boxDecoder guid
+            elif fullname = typeof<obj>.FullName then
+                value
+            else autoDecodeRecordsAndUnions t isCamelCase isOptional
 
     type Auto =
         static member generateDecoder<'T>(?isCamelCase : bool, [<Inject>] ?resolver: ITypeResolver<'T>): Decoder<'T> =
             let isCamelCase = defaultArg isCamelCase false
-            resolver.Value.ResolveType() |> (autoDecoder isCamelCase) |> unboxDecoder
+            resolver.Value.ResolveType() |> (autoDecoder isCamelCase false) |> unboxDecoder
 
         static member fromString<'T>(json: string, ?isCamelCase : bool, [<Inject>] ?resolver: ITypeResolver<'T>): Result<'T, string> =
             let decoder = Auto.generateDecoder(?isCamelCase=isCamelCase, ?resolver=resolver)